requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
dj-database-url
gunicorn
twilio